requirements.txt
cffi==1.14.3
chardet==3.0.4
coloredlogs==14.0
cryptography==3.1.1
humanfriendly==8.2
img2pdf==0.4.0
lxml==4.5.2
ocrmypdf==11.2.1
pdfminer.six==20200726
pikepdf==1.19.3
Pillow==7.2.0
pluggy==0.13.1
pycparser==2.20
PyYAML==5.3.1
reportlab==3.5.53
simple-term-menu==0.10.2
six==1.15.0
sortedcontainers==2.2.2
tqdm==4.50.2
